1.3.2 upstream specfile reconciliation:  %description and Group: tags
1.3.2 upstream specfile reconciliation:

Sol Jerome requested the differences between the upstream specfile in
misc/bcfg2.spec and the Fedora specfile be reconciled as much as
possible.  This will hopefully make updates easier for distro
packagers (like me).

The approach is to pick one feature at a time and massage both
upstream and Fedora versions so that they match as much as possible
after several easy-to-understand commits on each side.  Beneficial
present in one but not the other will be preserved in both.

Unfortunately, they will never match perfectly since upstream
maintains support for other RPM-based distros.  I believe Fedora
Packaging Guidelines does not permit (this statement should be
verified).  An additional goal is to make a diff between the two
minimal and easy to read.

------------------

Reconcile %description and Group: tags